fix regenerate script to fail if the regenerate_binary_html not present

+++ rebuild_html.sh
@@ -2,6 +2,8 @@
 cd src
 cargo build --release
 cd ..
-rm regenerate_html_bin
+if test -f "regenerate_html_bin"; then
+    rm regenerate_html_bin
+fi
 mv src/target/release/regenerate_html_bin regenerate_html_bin
 ./regenerate_html_bin
